refactor: extract some manual backport operations
